updated practice mode
done

+++ app/src/main/java/com/arhamsoft/deskilz/networking/networkModels/GetPlayerAccountData.kt
@@ -3,6 +3,6 @@ package com.arhamsoft.deskilz.networking.networkModels
 data class GetPlayerAccountData( //getcustomplayerdata and getgamecustomdata same model
     //    val userTickets: Int=0,
     val userDollers: Double=0.0,
-    val userDeskillzCurrency: Long=0,
+    val userDeskillzCurrency: Double=0.0,
     val busd: Double=0.0
 )
